requirements.txt
ai-services-migrate==0.1
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==2.4.1
asyncclick==8.3.0.7
asyncpg==0.31.0
bcrypt==5.0.0
build==1.4.2
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cleo==2.1.0
click==8.3.1
clickhouse-connect==0.15.1
crashtest==0.4.1
cryptography==46.0.7
decorator==5.2.1
devtools==0.12.2
dictdiffer==0.9.0
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
dulwich==1.1.0
ecdsa==0.19.2
email-validator==2.3.0
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.135.1
fastjsonschema==2.21.2
filelock==3.25.2
findpython==0.7.1
flatbuffers==25.12.19
fsspec==2026.2.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.11
ImageHash==4.3.2
installer==0.7.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
iso8601==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
keyring==25.7.0
lz4==4.4.5
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
nltk==3.9.4
numpy==2.4.3
onnx==1.21.0
onnxruntime==1.24.4
openai==2.31.0
optimum==2.1.0
optimum-onnx==0.1.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
passlib==1.7.4
pbs-installer==2026.3.25
pexpect==4.9.0
pillow==12.1.1
pkginfo==1.12.1.2
platformdirs==4.9.4
poetry-core==2.3.2
portalocker==3.2.0
prompt_toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.2
psycopg2==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pypika-tortoise==0.6.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-discovery==1.2.1
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pytz==2026.1.post1
PyWavelets==1.9.0
PyYAML==6.0.3
qdrant-client==1.17.1
RapidFuzz==3.14.3
redis==7.4.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rsa==4.9.1
ruff==0.15.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.14.0
torch==2.10.0
tortoise-orm==1.1.7
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
trove-classifiers==2026.1.14.14
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
virtualenv==21.2.0
wcwidth==0.6.0
xattr==1.3.0
zstandard==0.25.0
